+++ resources/views/event/admin/import.blade.php
@@ -39,23 +39,19 @@
 
                         <div class="form-group row">
 
-                            <label for="ctot" class="col-md-2 col-form-label text-md-right"> Format</label>
+                            <label for="ctot" class="col-md-2 col-form-label text-md-right"> Headers in <strong>bold</strong> are always required</label>
                             <div class="col-md-10">
                                 <div class="form-control-plaintext">
                                     @if($event->event_type_id == \App\Enums\EventType::MULTIFLIGHTS)
-                                        <abbr title="[hh:mm]">CTOT 1</abbr> - <abbr title="[ICAO]">Airport 1</abbr> -
-                                        <abbr title="[hh:mm]">CTOT 2</abbr> - <abbr title="[ICAO]">Airport 2</abbr> -
-                                        <abbr title="[ICAO]">Airport 3</abbr>
+                                        <strong><abbr title="[hh:mm]">CTOT 1</abbr></strong> - <strong><abbr title="[ICAO]">Airport 1</abbr></strong> -
+                                        <strong><abbr title="[hh:mm]">CTOT 2</abbr></strong> - <strong><abbr title="[ICAO]">Airport 2</abbr></strong> -
+                                        <strong><abbr title="[ICAO]">Airport 3</abbr></strong>
                                     @else
-                                        <strong>Arrivals</strong> - Call Sign | <abbr title="[ICAO]">Origin</abbr> |
-                                        <abbr title="[ICAO]">Destination</abbr> | <abbr title="[hh:mm]">ETA</abbr> |
-                                        <abbr title="[ICAO]">Aircraft</abbr>
-                                        Type
-                                        <br>
-                                        <strong>Departures</strong> - Call Sign | <abbr title="[ICAO]">Origin</abbr> |
-                                        <abbr title="[ICAO]">Destination</abbr> |
-                                        <abbr title="[hh:mm]">EOBT</abbr> | <abbr title="[ICAO]">Aircraft</abbr>
-                                        Type
+                                        Call Sign | <strong><abbr title="[ICAO]">Origin</abbr></strong> |
+                                        <strong><abbr title="[ICAO]">Destination</abbr></strong> |
+                                        <abbr title="[hh:mm]">EOBT</abbr> | <abbr title="[hh:mm]">ETA</abbr> |
+                                        <abbr title="[ICAO]">Aircraft Type</abbr> | Route
+
                                     @endif
                                 </div>
                             </div>